requirements-final.txt
# Core dependencies (pre-built wheels)
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pymongo==4.5.0
tqdm==4.66.1
retry==0.9.2
fake-useragent==1.2.1
PyPDF2==3.0.1
python-docx==0.8.11

# Data analysis
numpy==1.26.3
pandas==2.1.4

# Testing
pytest==7.4.0

# Use httpx instead of aiohttp (has pre-built wheels)
httpx==0.27.0